[CDNG-4276]: Add expression functor token in ambiance (#20451)

+++ 870-orchestration/src/main/java/io/harness/engine/expressions/functors/SecretFunctor.java
@@ -10,9 +10,9 @@ import lombok.Value;
 @OwnedBy(CDC)
 @Value
 public class SecretFunctor implements ExpressionFunctor {
-  int expressionFunctorToken;
+  long expressionFunctorToken;
 
-  public SecretFunctor(int expressionFunctorToken) {
+  public SecretFunctor(long expressionFunctorToken) {
     this.expressionFunctorToken = expressionFunctorToken;
   }
